Fix encodeJSON issues
Now rejects all characters and skips optional types that are not
currently present.

+++ kalix-example/src/main/java/eu/arrowhead/kalix/example/dto/Shape.java
@@ -13,7 +13,7 @@ public interface Shape {
     Point position();
     ShapeType type();
     Optional<String> name();
-    Map<String, Map<String, Short>> properties();
+    Map<String, Map<String, Integer>> properties();
     List<List<Byte>> attributes();
     int[] attributes2();
 }
